# swagger.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Micro Service Project - Coalition"
host: "coalition.com"
basePath: "/api/"
...
schemes:
  - "https"
  - "http"
paths:
...
  /servers:
    post:
      tags:
        - "Server"
      summary : "Create server"
      produces :
        - "application/json"
      parameters:
        - in: body
          name : "body"
          required: true
          schema:
            type: object
            properties: 
              ownerID:
                type: integer
              serverName:
                type: "string"
      responses:
        "201":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              userID:
                type: integer
        "404":
          description: "User not found"
        "500":
          description: "Internal server error"
    get:
      tags:
        - "Server"
      summary: "Get all servers"
      produces:
        - "application/json"
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                server_id: 
                  type : integer
                server_name:
                  type : string
                owner_id :
                  type : integer
                created_at :
                  type : string
                updated_at :
                  type: string
        "500":
          description: "Internal server error"
  /servers/{serverID}:
    parameters:
      - name: "serverID"
        in: path
        description: "ID of server"
        required: true
        type: integer
    get:
      tags:
        - "Server"
      summary: "Get all servers"
      produces:
        - "application/json"
      responses:
        "200":
          description:  "successful operation"
          schema:
            type: object
            properties:
              server_name:
                type : string
              owner_id :
                type : integer
              created_at :
                type : string
              updated_at :
                type: string
        "404":
          description:  "Server not found"
        "500":
          description: "Internal server error"
    put:
      tags:
        - "Server"
      summary: "Change server by ID"
      produces:
        - "application/json"
      parameters:
        - in: body
          name : "body"
          required: true
          schema:
            type: object
            properties: 
              userID:
                type: integer
              serverName:
                type: string
      responses:
        "200":
          description:  "successful operation"
        "403":
          description:  "Access forbidden"
        "404":
          description:  "Server or User not found"
        "500":
          description: "Internal server error"
    delete:
      tags:
        - "Server"
      summary: "Delete server by ID"
      produces:
        - "application/json"
      parameters:
        - in: body
          name : "body"
          required: true
          schema:
            type: object
            properties: 
              userID:
                type: integer
      responses:
        "204":
          description:  "successful operation"
        "403":
          description:  "Access forbidden"
        "404":
          description:  "Server or User not found"
        "500":
          description: "Internal server error"
...
  /servers/members/server/{serverID} :
    get :
      tags:
      - "Server/Member"
      summary: "Get all members of a server"
      produces:
        - "application/json"
      parameters:
      - name: "serverID"
        in: path
        description: "ID of server"
        required: true
        type: integer
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                user_id :
                  type : integer
                created_at :
                  type : string
        "404":
          description: "Server not found"
        "500":
          description: "Internal server error"
...
  /servers/channels/server/{serverID} :
    get :
      tags:
      - "Server/Channel"
      summary: "Get all channels of a server"
      produces:
        - "application/json"
      parameters:
      - in: path
        name: "serverID"
        description: "ID of server"
        required: true
        type: integer
      - in : body
        name : body
        required: true
        schema:
          type: object
          properties: 
            userID:
              type: integer
      responses:
        "200":
          description:  "successful operation"
          schema:
            type : array
            items:
              type: object
              properties:
                conv_id :
                  type: integer
                channel_name :
                  type: string
                created_at :
                  type: string
                updated_at :
                  type: string
        "403":
          description: "Permission denied"
        "404":
          description: "Server or User not found"
        "500":
          description: "Internal server error"
...